Fix single charge basis transmon. Now using the standard truncation scheme, but keeping N_max_charge as the variable name...
