+++ qcsys/devices/__init__.py
@@ -9,7 +9,6 @@ from .kno import *
 from .transmon import *
 from .tunable_transmon import *
 from .transmon_single_charge_basis import *
-from .squid_transmon import *
 from .fluxonium import *
 from .ats import *
 from .ideal_qubit import *
